pyproject.toml
[project]
name = "ritm-ia"
version = "0.1.0"
description = "Schoolar project about pretiction musical tracks"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.4",
    "openai>=2.30.0",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
